removed FullNews component, used newsDetails(), changed query parameter to ?qInTitle="${title}", changed function name in api to getArticleDetails(), id => id, fromToId => fromStringToId, fromToStr => fromIdToString, changed article details loading
