--- src/components/FullNews/index.js
@@ -1,1 +1,0 @@
-export * from './FullNews'
