+++ src/components/Home/Home.jsx
@@ -17,6 +17,9 @@ import Typography from '@material-ui/core/Typography'
 // react-router-dom
 import { Link } from 'react-router-dom'
 
+// routes
+import { routes } from '../../config/routes'
+
 // styles
 import { useStyles } from './styles'
 
@@ -26,8 +29,8 @@ import { getNews } from '../../store/actions'
 // selectors
 import * as selectors from '../../store/selectors'
 
-// base64
-import { fromToId } from '../../utils/transformBase64'
+// base 64
+import { fromStringToId } from '../../utils/transformBase64/transformStringToId'
 
 const Home = () => {
   const entities = useSelector(selectors.getEntities('news'))
@@ -47,7 +50,7 @@ const Home = () => {
 
   const loadMore = () => {
     const nextPage = page + 1
-    dispatch(getNews(nextPage))
+    dispatch(getNews('us', nextPage))
   }
 
   return (
@@ -59,11 +62,7 @@ const Home = () => {
             <Grid key={item.publishedAt} item xs={12} sm={6} md={3}>
               <Card>
                 <CardHeader className={classes.author} subheader={item.author} />
-                <CardHeader
-                  className={classes.title}
-                  title="Shrimp and Chorizo Paella"
-                  subheader="September 14, 2016"
-                />
+                <CardHeader className={classes.title} title={item.title} subheader="September 14, 2016" />
                 <CardMedia
                   className={classes.media}
                   image={
@@ -78,7 +77,7 @@ const Home = () => {
                   </Typography>
                 </CardContent>
                 <div className={classes.link}>
-                  <Link to={`article/${fromToId(item.title)}`}>Read more...</Link>
+                  <Link to={routes.newsDetails(fromStringToId(item.title))}>Read more...</Link>
                 </div>
               </Card>
             </Grid>
