--- src/utils/transformBase64.js
@@ -1,2 +1,0 @@
-export const fromToId = str => btoa(unescape(encodeURIComponent(str)))
-export const fromToStr = str => decodeURIComponent(escape(window.atob(str)))
