+++ src/utils/transformBase64/transformIdToString.js
@@ -1,0 +1,1 @@
+export const fromIdToString = str => decodeURIComponent(escape(window.atob(str)))
