+++ src/utils/transformBase64/transformStringToId.jsx
@@ -1,0 +1,1 @@
+export const fromStringToId = str => btoa(unescape(encodeURIComponent(str)))
